refactor: modularize DataGrid components
